fix(interest): kyc card shows
